# oas/odin-schema.yml
openapi: '3.0.2'
info:
  title: API Title
  version: '0.0.1'
servers:
  - url: http://localhost:8080/api/v1
paths:
  /admin/token/:
    post:
      summary: Get token
      security: []
      description: Get token
      operationId: getToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - token
                properties:
                  token:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/token/:
    get:
      summary: Generate user token
      description: Generate user token
      operationId: generateUserToken
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - token
                properties:
                  token:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /executions/execute/:
    post:
      summary: Execute a script
      description: Execute a script
      operationId: execute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - jobId
                  - events
                properties:
                  jobId:
                    type: integer
                    format: int64
                  events:
                    type: string
                    format: url
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /jobs/execution/:
    get:
      summary: Get all execution jobs
      description: Get all execution jobs
      operationId: getAllExecutionJobs
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - jobs
                  - pagination
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /executions/:
    get:
      summary: Get all executions
      description: Get all executions
      operationId: getAllExecutions
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - executions
                  - pagination
                properties:
                  executions:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExecutionResult'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /jobs/{JobId}/executions/:
    get:
      summary: Get executions of given job
      description: Get executions of given job
      operationId: getExecutionsForJob
      parameters:
        - name: JobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - executions
                  - pagination
                properties:
                  executions:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExecutionResult'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execution job not found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /executions/{execId}/:
    get:
      summary: Get execution result by id
      description: Get execution result by id
      operationId: getExecutionResultById
      parameters:
        - name: execId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResult'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execution result not found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /executions/jobs/{JobId}/:
    get:
      summary: Get execution job
      description: Get execution job
      operationId: getExecutionJobById
      parameters:
        - name: JobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execution job not found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Cancel Execution Job
      description: Cancel Execution Job
      operationId: cancelExecutionJob
      parameters:
        - name: JobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete execution job
      description: Delete execution job
      operationId: deleteExecutionJob
      parameters:
        - name: JobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execution job not found
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /executions/workers:
    get:
      summary: Get all execution workers
      description: Get all execution workers
      operationId: getExecutionWorkers
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - workers
                  - pagination
                properties:
                  workers:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExecutionWorker'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /executions/workers/{workerId}/:
    delete:
      summary: Delete execution worker
      description: Delete execution worker
      operationId: deleteExecutionWorker
      parameters:
        - name: workerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: force
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execution worker not found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /version/:
    get:
      summary: Get version
      description: Get version
      operationId: getVersion
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - version
                properties:
                  version:
                    type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    PageNumber:
      name: page
      in: query
      description: The page number to retrieve.
      required: false
      schema:
        type: integer
        format: int32
        default: 0
    PageSize:
      name: pageSize
      in: query
      description: The number of items per page.
      required: false
      schema:
        type: integer
        format: int32
        default: 20
  schemas:
...
    ExecutionRequest:
      type: object
      required:
        - code
        - language
      properties:
        environment:
          oneOf:
            - $ref: '#/components/schemas/Flake'
            - $ref: '#/components/schemas/ExecutionEnvironmentSpec'
        config:
          oneOf:
            - $ref: '#/components/schemas/DockerExecutionConfig'
            - $ref: '#/components/schemas/PodmanExecutionConfig'
        code:
          type: string
        language:
          type: string
        max_retries:
          type: integer
          format: int4
          default: 5
        timeout:
          type: integer
          format: int32
          default: -1
        priority:
          type: integer
          format: int4
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    ExecutionWorker:
      type: object
      required:
        - id
        - name
        - status
        - created_at
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        created_at:
          type: string
          format: date-time
        status:
          type: string
        updated_at:
          type: string
          format: date-time
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []